Moved DiscordSnapshotMemberUpdateInfo to own file
